% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{edge_list_fn}
\alias{edge_list_fn}
\title{Edge list}
\usage{
edge_list_fn(A)
}
\arguments{
\item{A}{unweighted adjacency matrix of graph}
}
\value{
ordered list of edges
}
\description{
Edge list
}
\examples{
a <- matrix(c(0,1,1,0), nrow =2)
edge_list_fn(a)

}
