% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{triangle_list_fn}
\alias{triangle_list_fn}
\title{Triangle list}
\usage{
triangle_list_fn(A)
}
\arguments{
\item{A}{unweighted adjacency matrix}
}
\value{
ordered list of triangles
}
\description{
Triangle list
}
\examples{
a <- matrix(c(0,1,1,1,0,1,1,1,0), nrow = 3)
triangle_list_fn(a)

}
